Make sure we don't stop/start jail unnecessarily
This commit makes sure that we don't stop/start a jail unnecessarily when installing a package. It is only stopped if it was started by iocage during pkg installation phase.
Also this commit removes a usage where we removed resolv.conf, i am unsure why we were doing that here at all but that is likely going to cause issues if the jail is already running and the user expects his resolv.conf to be there.