Inherit mount point from parent
This commit fixes an issue where we set the mount point as that of the parent if iocage's mount point didn't start with it. But if altroots are involved, they are added automatically to the mount point resulting in a misconfigured mount point.

(cherry picked from commit a85d77271b06437eadf8d94d14169d65cb96bdc3)