Update plugin artifact before executing pre update script
This commit adds changes to upgrade implementation of plugin to make sure we update plugin artifact before trying to upgrade. Motivation behind this change is that we are going to try to execute pre update hook but without updating the artifact from upstream, we won't have the latest changes upstream wanted resulting in a potential inconsistent upgrade.